Rename to display name of component
