+++ app/detail/Pane.tsx
@@ -4,7 +4,7 @@ import {SuricataEvent} from "ppl/detail/models/SuricataEvent"
 import {ZeekEvent} from "ppl/detail/models/ZeekEvent"
 import RelatedAlerts from "ppl/detail/RelatedAlerts"
 import RelatedConns from "ppl/detail/RelatedConns"
-import CorrelationPanel from "ppl/detail/Correlation"
+import UidPanel from "ppl/detail/UidPanel"
 import React, {useLayoutEffect, memo, useMemo, useRef} from "react"
 import {useSelector} from "react-redux"
 import ConnPanel from "src/js/components/LogDetails/ConnPanel"
@@ -41,7 +41,7 @@ const Content = memo<Props>(function Content({record}) {
         <Fields record={record} />
       </div>
       <div className="column">
-        {isZeek && uid && <CorrelationPanel record={record} />}
+        {isZeek && uid && <UidPanel record={record} />}
         {isSuricata && cid && <RelatedAlerts record={record} />}
         {isSuricata && cid && <RelatedConns record={record} />}
         {isConn && <ConnPanel record={record} />}
