+++ ppl/detail/UidPanel.tsx
